email-processing-just-code/requirements_api.txt
fastapi
uvicorn
sqlalchemy>=2.0
alembic
psycopg2-binary
python-dotenv
requests
msal
beautifulsoup4
